Wait for the Nix store to mount
Almost every time I restart my computer, dmn fails to run until I logout
and log back in. This is because the Nix store isn't mounted when the
agent tries to start, so it fails enough times that launchd puts it in
the penalty box. Waiting for the Nix store avoids this issue.

Be aware that this will cause the entry in System Settings to change
from `dmn` to `sh`. Ideally this will be addressed in the future, but I
am considering it a worthwhile tradeoff to avoid having to log out and
back in on every restart.